requirements.txt
msgpack
msgpack-numpy
dm-haiku
jax
optax
rich
polars